+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { Code } from './Code';
 export { ContentEditable } from './ContentEditable';
 export { History } from './History';
 export { Selection } from './Selection';