+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.10.0 2022-06-01
+
+- [FEATURE] Basic code block only supporing elixir synthax highlighting
+- [FIX] Using button to add block after a code or table block fails
+
 # 0.9.4 2022-05-30
 
 - [FIX] Bug with typing after inserting new line at the end of block